+++ src/subcommands/tags.rs
@@ -1,32 +1,37 @@
 use crate::note::Note;
 use anyhow::Result;
-use std::collections::BTreeSet;
-use std::ffi::OsStr;
-use std::fs;
+use ignore::{overrides::OverrideBuilder, types::TypesBuilder, WalkBuilder};
+use std::collections::HashSet;
 
 pub fn run() -> Result<()> {
-    let mut tags: BTreeSet<String> = BTreeSet::new();
+    let markdown_matcher = TypesBuilder::new()
+        .add_defaults()
+        .select("markdown")
+        .build()?;
 
-    for entry in fs::read_dir(".")? {
-        let p = &entry?.path();
+    // This override, which ignores hidden entries, is necessary because the markdown matcher
+    // itself overrides the WalkBuilder's default filtering of hidden entries.
+    let hidden_override = OverrideBuilder::new("./").add("!.*")?.build()?;
 
-        if p.is_dir() {
-            continue;
-        }
+    let entries: Vec<_> = WalkBuilder::new("./")
+        .types(markdown_matcher)
+        .overrides(hidden_override)
+        .build()
+        .into_iter()
+        .filter_map(|r| r.ok())
+        .collect();
 
-        if p.extension().unwrap_or_default() != OsStr::new("md") {
-            continue;
-        }
+    let mut tags: Vec<_> = entries
+        .iter()
+        .map(|entry| match Note::read_from_file(entry.path()) {
+            Ok(n) => n.front_matter.tags,
+            Err(_) => HashSet::new(),
+        })
+        .flatten()
+        .collect();
 
-        match Note::read_from_file(&p) {
-            Ok(n) => {
-                for tag in n.front_matter.tags {
-                    tags.insert(tag);
-                }
-            }
-            Err(_) => continue,
-        };
-    }
+    tags.sort();
+    tags.dedup();
 
     for t in tags {
         println!("{}", t);