+++ tests/subcommands/test_tags.rs
@@ -23,7 +23,7 @@ links: []
     );
     let content2 = format!(
         "---
-title: A
+title: B
 created: \"2020-04-19T18:23:24.774140Z\"
 tags:
   - {}
@@ -55,7 +55,7 @@ links: []
     );
     let content2 = format!(
         "---
-title: A
+title: B
 created: \"2020-04-19T18:23:24.774140Z\"
 tags:
   - {}
@@ -66,6 +66,31 @@ links: []
     t.setup_fs(vec![("a.md", &content1), ("b.md", &content2)])?;
     let output = t.zeke_tags()?.output()?;
     let stdout = str::from_utf8(&output.stdout)?;
+    assert_eq!(stdout, format!("{}\n", tag1));
+    Ok(())
+}
+
+#[test]
+fn reads_files_in_subdirectories() -> Result<()> {
+    let t = ZekeTester::new();
+    let tag1 = "red";
+    let content1 = format!(
+        "---
+title: A
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag1
+    );
+
+    t.temp_dir.child("subdir").create_dir_all()?;
+    t.setup_fs(vec![("subdir/a.md", &content1)])?;
+    let output = t.zeke_tags()?.output()?;
+    let stdout = str::from_utf8(&output.stdout)?;
+    let test = t.temp_dir.into_persistent();
+    println!("{:?}", test.path().display());
     assert_eq!(stdout, format!("{}\n", tag1));
     Ok(())
 }
@@ -85,7 +110,7 @@ links: []
         tag1
     );
     let content2 = "---
-title: A
+title: B
 created: \"2020-04-19T18:23:24.774140Z\"
 tags:
   - green
@@ -96,6 +121,153 @@ links: []
     let output = t.zeke_tags()?.output()?;
     let stdout = str::from_utf8(&output.stdout)?;
     assert_eq!(stdout, format!("{}\n", tag1));
+    Ok(())
+}
+
+#[test]
+fn does_not_read_hidden_files() -> Result<()> {
+    let t = ZekeTester::new();
+    let tag1 = "red";
+    let content1 = format!(
+        "---
+title: A
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag1
+    );
+    let content2 = "---
+title: B
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - green
+  - blue
+links: []
+--- ";
+    t.setup_fs(vec![("a.md", &content1), (".hidden.md", &content2)])?;
+
+    let output = t.zeke_tags()?.output()?;
+    let stdout = str::from_utf8(&output.stdout)?;
+
+    assert_eq!(stdout, format!("{}\n", tag1));
+    Ok(())
+}
+
+#[test]
+fn does_not_read_files_specified_via_ignore() -> Result<()> {
+    let t = ZekeTester::new();
+    let tag1 = "red";
+    let content1 = format!(
+        "---
+title: A
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag1
+    );
+    let path2 = "b.md";
+    let content2 = "---
+title: B
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - green
+  - blue
+links: []
+--- ";
+    let ignore_content = format!("{}", path2);
+    t.setup_fs(vec![
+        ("a.md", &content1),
+        (path2, &content2),
+        (".ignore", &ignore_content),
+    ])?;
+
+    let output = t.zeke_tags()?.output()?;
+    let stdout = str::from_utf8(&output.stdout)?;
+
+    assert_eq!(stdout, format!("{}\n", tag1));
+    Ok(())
+}
+
+#[test]
+fn does_not_read_files_specified_via_gitignore_if_git() -> Result<()> {
+    let t = ZekeTester::new();
+    let tag1 = "red";
+    let content1 = format!(
+        "---
+title: A
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag1
+    );
+    let path2 = "b.md";
+    let content2 = "---
+title: B
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - green
+  - blue
+links: []
+--- ";
+    let ignore_content = format!("{}", path2);
+    t.setup_fs(vec![
+        ("a.md", &content1),
+        (path2, &content2),
+        (".gitignore", &ignore_content),
+    ])?;
+    // This is done to fake that we're in a git repo
+    t.temp_dir.child(".git").create_dir_all()?;
+
+    let output = t.zeke_tags()?.output()?;
+    let stdout = str::from_utf8(&output.stdout)?;
+
+    assert_eq!(stdout, format!("{}\n", tag1));
+    Ok(())
+}
+
+#[test]
+fn reads_files_specified_via_gitignore_if_no_git() -> Result<()> {
+    let t = ZekeTester::new();
+    let tag1 = "red";
+    let tag2 = "blue";
+    let content1 = format!(
+        "---
+title: A
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag1
+    );
+    let path2 = "b.md";
+    let content2 = format!(
+        "---
+title: B
+created: \"2020-04-19T18:23:24.774140Z\"
+tags:
+  - {}
+links: []
+--- ",
+        tag2
+    );
+    let ignore_content = format!("{}", path2);
+    t.setup_fs(vec![
+        ("a.md", &content1),
+        (path2, &content2),
+        (".gitignore", &ignore_content),
+    ])?;
+
+    let output = t.zeke_tags()?.output()?;
+    let stdout = str::from_utf8(&output.stdout)?;
+
+    assert_eq!(stdout, format!("{}\n{}\n", tag2, tag1));
     Ok(())
 }